More tests for Vec<Accessor>
